Fix bug with tightly packed data (#16)
There's an issue with our buffering of input data where we may consume
more data than is in the input stream (not a buffer overflow). This is
particularly an issue if the caller has data tightly packed data with
streams one after the other, but not separated in the input buffer (e.g.
they just read data from a file & don't know how long each stream is).

The fix is to "put back" data that we didn't use, but ONLY in situations
where we didn't terminate due to not having enough data. This allows us
to continue to function when called with only one byte of input at a
time.